refactor(#4712): de-giant manual_rebind/mod.rs — pure-move test-barrier cluster to test_barriers.rs
Extract the test-only race-seam barrier cluster (PostAdoptionClaimBarrier /
EpisodeAuthorityHeldBarrier types, install slots, restore-on-drop guards,
install_* and await_* hooks) verbatim from manual_rebind/mod.rs into a new
manual_rebind/test_barriers.rs submodule. Zero logic change: every item keeps
its per-item #[cfg(test)] gate (no new logical test module, test-lane baseline
untouched); the two formerly-private await_* hooks are elevated to pub(super)
as the minimum surface for the parent and episode_handoff.rs to keep calling
them; the four pub(crate) names are chain re-exported by mod.rs so the
recovery_engine root re-export surface is untouched.

mod.rs: 1001 -> 911 prod LoC (below the 1000 giant threshold);
test_barriers.rs: 114 prod LoC. Remove the manual_rebind/mod.rs shrink entry
from scripts/giant_file_registry.toml (ghost-registration rule), clear the
frozen-giant-surface marker in change-surfaces.md, and regenerate inventory
docs (ARCHITECTURE.md tree sync).

Module tests before == after: 55 passed; 0 failed; 6858 filtered out.
